fix equals method in chain; remove idPicture (also in schema); modify UsersDAOImp & GroupDAOImpl upon previous fix.

+++ src/main/java/com/louay/projects/model/dao/InsertGroupPostDAO.java
@@ -7,5 +7,5 @@ public interface InsertGroupPostDAO {
 
     Long insertGroupComments(GroupComments comment);
 
-    Long insertGroupPicture(GroupPicture picture);
+    int insertGroupPicture(GroupPicture picture);
 }
